refactor(engine): split cheap canonicalization into fold pass
  Extract local, schema-free, idempotent rewrites from Simplify into a
  new `engine::fold` pass. Fold runs bottom-up before each Simplify
  visit so heavyweight rules see canonical input.

  Migrated to fold: expr_and, expr_or, expr_binary_op, expr_cast,
  expr_in_list, expr_is_null, expr_list, expr_match, expr_not,
  expr_record. Heavyweight rules (predicate inference, structural
  rewrites needing schema/eval) remain in simplify.

  expr_project stays in simplify because its Match-distribution rule
  relies on the dead-arm protection in `visit_expr_match_mut` to avoid
  panics on unreachable invalid projections.

  Test layout follows: fold/tests/* covers fold rules in isolation;
  simplify/tests/* keeps end-to-end visit_expr_mut tests. Adds an
  idempotence proptest for fold (`fold(fold(x)) == fold(x)`).

  284 lib tests pass, clippy clean, fmt applied. The doctest failure in
  db::builder::Builder::connect is pre-existing on origin/main (unrelated to
  fold work).
