+++ crates/toasty/src/engine/fold/tests.rs
@@ -1,0 +1,11 @@
+mod expr_and;
+mod expr_binary_op;
+mod expr_cast;
+mod expr_in_list;
+mod expr_is_null;
+mod expr_list;
+mod expr_match;
+mod expr_not;
+mod expr_or;
+mod expr_record;
+mod prop_idempotence;
