+++ crates/toasty/src/engine/fold/tests/expr_in_list.rs
@@ -1,5 +1,4 @@
-use super::test_schema;
-use crate::engine::simplify::Simplify;
+use crate::engine::fold::expr_in_list::fold_expr_in_list;
 use toasty_core::stmt::{self, Expr, Value};
 
 /// Helper to construct an "in list" expression.
@@ -22,12 +21,9 @@ fn expr_list(items: Vec<Expr>) -> Expr {
 
 #[test]
 fn empty_value_list_becomes_false() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), []) → false`
     let mut expr = in_list(Expr::arg(0), value_list(vec![]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_false());
@@ -35,12 +31,9 @@ fn empty_value_list_becomes_false() {
 
 #[test]
 fn empty_expr_list_becomes_false() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), list([])) → false`
     let mut expr = in_list(Expr::arg(0), expr_list(vec![]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_false());
@@ -48,12 +41,9 @@ fn empty_expr_list_becomes_false() {
 
 #[test]
 fn single_value_becomes_eq() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), [42]) → eq(arg(0), 42)`
     let mut expr = in_list(Expr::arg(0), value_list(vec![Value::from(42i64)]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(matches!(
         result,
@@ -64,12 +54,9 @@ fn single_value_becomes_eq() {
 
 #[test]
 fn single_expr_becomes_eq() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), list([arg(1)])) → eq(arg(0), arg(1))`
     let mut expr = in_list(Expr::arg(0), expr_list(vec![Expr::arg(1)]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(matches!(
         result,
@@ -80,39 +67,30 @@ fn single_expr_becomes_eq() {
 
 #[test]
 fn two_values_unchanged() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), [1, 2])`, multiple items, not simplified
     let mut expr = in_list(
         Expr::arg(0),
         value_list(vec![Value::from(1i64), Value::from(2i64)]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
 }
 
 #[test]
 fn two_exprs_unchanged() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), list([arg(1), arg(2)]))`, multiple items, not simplified
     let mut expr = in_list(Expr::arg(0), expr_list(vec![Expr::arg(1), Expr::arg(2)]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
 }
 
 #[test]
 fn arg_in_single() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), [42]) → eq(arg(0), 42)`
     let mut expr = in_list(Expr::arg(0), value_list(vec![Value::from(42i64)]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(matches!(
         result,
@@ -123,12 +101,9 @@ fn arg_in_single() {
 
 #[test]
 fn arg_in_empty() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), []) → false`
     let mut expr = in_list(Expr::arg(0), value_list(vec![]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_false());
@@ -136,15 +111,12 @@ fn arg_in_empty() {
 
 #[test]
 fn arg_in_multi() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `in_list(arg(0), [1, 2])`, multiple items, not simplified
     let mut expr = in_list(
         Expr::arg(0),
         value_list(vec![Value::from(1i64), Value::from(2i64)]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
 }
@@ -153,9 +125,6 @@ fn arg_in_multi() {
 
 #[test]
 fn null_in_list_becomes_null() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `null in (1, 2, 3)` → `null`
     let mut expr = in_list(
         Expr::null(),
@@ -165,7 +134,7 @@ fn null_in_list_becomes_null() {
             Value::from(3i64),
         ]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_value_null());
@@ -173,12 +142,9 @@ fn null_in_list_becomes_null() {
 
 #[test]
 fn null_in_single_item_becomes_null() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `null in (42)` → `null`
     let mut expr = in_list(Expr::null(), value_list(vec![Value::from(42i64)]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_value_null());
@@ -186,12 +152,9 @@ fn null_in_single_item_becomes_null() {
 
 #[test]
 fn null_in_expr_list_becomes_null() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `null in list([arg(0), arg(1)])` → `null`
     let mut expr = in_list(Expr::null(), expr_list(vec![Expr::arg(0), Expr::arg(1)]));
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_value_null());
@@ -201,9 +164,6 @@ fn null_in_expr_list_becomes_null() {
 
 #[test]
 fn dedup_all_identical_values() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `x in (1, 1, 1)` → `x = 1`
     let mut expr = in_list(
         Expr::arg(0),
@@ -213,7 +173,7 @@ fn dedup_all_identical_values() {
             Value::from(1i64),
         ]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     // Dedup leaves one item, which is then rewritten to equality
     assert!(matches!(result, Some(Expr::BinaryOp(_))));
@@ -221,9 +181,6 @@ fn dedup_all_identical_values() {
 
 #[test]
 fn dedup_leading_duplicate() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `x in (1, 1, 2)` → `x in (1, 2)`
     let mut expr = in_list(
         Expr::arg(0),
@@ -233,7 +190,7 @@ fn dedup_leading_duplicate() {
             Value::from(2i64),
         ]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
     let Expr::Value(Value::List(values)) = &*expr.list else {
@@ -246,9 +203,6 @@ fn dedup_leading_duplicate() {
 
 #[test]
 fn dedup_trailing_duplicate() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `x in (1, 2, 2)` → `x in (1, 2)`
     let mut expr = in_list(
         Expr::arg(0),
@@ -258,7 +212,7 @@ fn dedup_trailing_duplicate() {
             Value::from(2i64),
         ]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
     let Expr::Value(Value::List(values)) = &*expr.list else {
@@ -271,9 +225,6 @@ fn dedup_trailing_duplicate() {
 
 #[test]
 fn dedup_preserves_order() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `x in (3, 1, 2, 1, 3)` → `x in (3, 1, 2)` — first occurrence wins
     let mut expr = in_list(
         Expr::arg(0),
@@ -285,7 +236,7 @@ fn dedup_preserves_order() {
             Value::from(3i64),
         ]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
     let Expr::Value(Value::List(values)) = &*expr.list else {
@@ -299,9 +250,6 @@ fn dedup_preserves_order() {
 
 #[test]
 fn dedup_no_duplicates_unchanged() {
-    let schema = test_schema();
-    let simplify = Simplify::new(&schema);
-
     // `x in (1, 2, 3)` — no duplicates, list unchanged
     let mut expr = in_list(
         Expr::arg(0),
@@ -311,7 +259,7 @@ fn dedup_no_duplicates_unchanged() {
             Value::from(3i64),
         ]),
     );
-    let result = simplify.simplify_expr_in_list(&mut expr);
+    let result = fold_expr_in_list(&mut expr);
 
     assert!(result.is_none());
     let Expr::Value(Value::List(values)) = &*expr.list else {
