+++ crates/toasty/src/engine/fold/tests/expr_list.rs
@@ -1,12 +1,8 @@
-use super::test_schema;
-use crate::engine::simplify::Simplify;
+use crate::engine::fold::expr_list::fold_expr_list;
 use toasty_core::stmt::{Expr, ExprList, Value};
 
 #[test]
 fn all_const_values_become_value_list() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `list([1, 2, 3]) → [1, 2, 3]`
     let mut expr = ExprList {
         items: vec![
@@ -15,7 +11,7 @@ fn all_const_values_become_value_list() {
             Expr::Value(Value::from(3i64)),
         ],
     };
-    let result = simplify.simplify_expr_list(&mut expr);
+    let result = fold_expr_list(&mut expr);
 
     assert!(result.is_some());
     let Expr::Value(Value::List(list)) = result.unwrap() else {
@@ -29,9 +25,6 @@ fn all_const_values_become_value_list() {
 
 #[test]
 fn mixed_types_in_list() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `list(["hello", 42, true]) → ["hello", 42, true]`
     let mut expr = ExprList {
         items: vec![
@@ -40,7 +33,7 @@ fn mixed_types_in_list() {
             Expr::Value(Value::from(true)),
         ],
     };
-    let result = simplify.simplify_expr_list(&mut expr);
+    let result = fold_expr_list(&mut expr);
 
     assert!(result.is_some());
     let Expr::Value(Value::List(list)) = result.unwrap() else {
@@ -54,12 +47,9 @@ fn mixed_types_in_list() {
 
 #[test]
 fn empty_list_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `list([]) → []`
     let mut expr = ExprList { items: vec![] };
-    let result = simplify.simplify_expr_list(&mut expr);
+    let result = fold_expr_list(&mut expr);
 
     assert!(result.is_some());
     let Expr::Value(Value::List(list)) = result.unwrap() else {
@@ -70,28 +60,22 @@ fn empty_list_simplified() {
 
 #[test]
 fn non_const_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `list([1, arg(0)])`, non-constant, not simplified
     let mut expr = ExprList {
         items: vec![Expr::Value(Value::from(1i64)), Expr::arg(0)],
     };
-    let result = simplify.simplify_expr_list(&mut expr);
+    let result = fold_expr_list(&mut expr);
 
     assert!(result.is_none());
 }
 
 #[test]
 fn list_with_error_item_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `list([1, error("boom")])` — error is not a Value, so not folded
     let mut expr = ExprList {
         items: vec![Expr::Value(Value::from(1i64)), Expr::error("boom")],
     };
-    let result = simplify.simplify_expr_list(&mut expr);
+    let result = fold_expr_list(&mut expr);
 
     assert!(result.is_none());
 }
