+++ crates/toasty/src/engine/fold/tests/expr_not.rs
@@ -1,5 +1,4 @@
-use super::test_schema;
-use crate::engine::simplify::Simplify;
+use crate::engine::fold::expr_not::fold_expr_not;
 use toasty_core::stmt::{BinaryOp, Expr, ExprBinaryOp, ExprNot, Value};
 
 fn not_expr(expr: Expr) -> ExprNot {
@@ -12,28 +11,22 @@ fn not_expr(expr: Expr) -> ExprNot {
 
 #[test]
 fn double_negation_eliminated() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(not(true))` → `true`
     let inner = Expr::not(Expr::Value(Value::Bool(true)));
     let mut outer = not_expr(inner);
 
-    let result = simplify.simplify_expr_not(&mut outer);
+    let result = fold_expr_not(&mut outer);
 
     assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
 }
 
 #[test]
 fn triple_negation_reduces_to_single() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(not(not(true)))` → `not(true)`
     let inner = Expr::not(Expr::not(Expr::Value(Value::Bool(true))));
     let mut outer = not_expr(inner);
 
-    let result = simplify.simplify_expr_not(&mut outer);
+    let result = fold_expr_not(&mut outer);
     assert!(matches!(result, Some(Expr::Not(_))));
 }
 
@@ -41,39 +34,30 @@ fn triple_negation_reduces_to_single() {
 
 #[test]
 fn not_true_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(true)` → `false`
     let mut expr = not_expr(Expr::Value(Value::Bool(true)));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
 }
 
 #[test]
 fn not_false_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(false)` → `true`
     let mut expr = not_expr(Expr::Value(Value::Bool(false)));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
 }
 
 #[test]
 fn not_null_becomes_null() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(null)` → `null`
     let mut expr = not_expr(Expr::null());
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     assert!(matches!(result, Some(Expr::Value(Value::Null))));
 }
@@ -82,9 +66,6 @@ fn not_null_becomes_null() {
 
 #[test]
 fn not_eq_becomes_ne() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(1 = 2)` → `1 != 2`
     let mut expr = not_expr(Expr::BinaryOp(ExprBinaryOp {
         lhs: Box::new(Expr::Value(Value::from(1i64))),
@@ -92,7 +73,7 @@ fn not_eq_becomes_ne() {
         rhs: Box::new(Expr::Value(Value::from(2i64))),
     }));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::BinaryOp(binary_op)) = result else {
         panic!("expected `BinaryOp`");
@@ -104,9 +85,6 @@ fn not_eq_becomes_ne() {
 
 #[test]
 fn not_ne_becomes_eq() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(1 != 2)` → `1 = 2`
     let mut expr = not_expr(Expr::BinaryOp(ExprBinaryOp {
         lhs: Box::new(Expr::Value(Value::from(1i64))),
@@ -114,7 +92,7 @@ fn not_ne_becomes_eq() {
         rhs: Box::new(Expr::Value(Value::from(2i64))),
     }));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::BinaryOp(binary_op)) = result else {
         panic!("expected `BinaryOp`");
@@ -126,9 +104,6 @@ fn not_ne_becomes_eq() {
 
 #[test]
 fn not_lt_becomes_ge() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(1 < 2)` → `1 >= 2`
     let mut expr = not_expr(Expr::BinaryOp(ExprBinaryOp {
         lhs: Box::new(Expr::Value(Value::from(1i64))),
@@ -136,7 +111,7 @@ fn not_lt_becomes_ge() {
         rhs: Box::new(Expr::Value(Value::from(2i64))),
     }));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::BinaryOp(binary_op)) = result else {
         panic!("expected `BinaryOp`");
@@ -148,9 +123,6 @@ fn not_lt_becomes_ge() {
 
 #[test]
 fn not_ge_becomes_lt() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(1 >= 2)` → `1 < 2`
     let mut expr = not_expr(Expr::BinaryOp(ExprBinaryOp {
         lhs: Box::new(Expr::Value(Value::from(1i64))),
@@ -158,7 +130,7 @@ fn not_ge_becomes_lt() {
         rhs: Box::new(Expr::Value(Value::from(2i64))),
     }));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::BinaryOp(binary_op)) = result else {
         panic!("expected `BinaryOp`");
@@ -170,9 +142,6 @@ fn not_ge_becomes_lt() {
 
 #[test]
 fn not_gt_becomes_le() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(1 > 2)` → `1 <= 2`
     let mut expr = not_expr(Expr::BinaryOp(ExprBinaryOp {
         lhs: Box::new(Expr::Value(Value::from(1i64))),
@@ -180,7 +149,7 @@ fn not_gt_becomes_le() {
         rhs: Box::new(Expr::Value(Value::from(2i64))),
     }));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::BinaryOp(binary_op)) = result else {
         panic!("expected `BinaryOp`");
@@ -192,9 +161,6 @@ fn not_gt_becomes_le() {
 
 #[test]
 fn not_le_becomes_gt() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(1 <= 2)` → `1 > 2`
     let mut expr = not_expr(Expr::BinaryOp(ExprBinaryOp {
         lhs: Box::new(Expr::Value(Value::from(1i64))),
@@ -202,7 +168,7 @@ fn not_le_becomes_gt() {
         rhs: Box::new(Expr::Value(Value::from(2i64))),
     }));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::BinaryOp(binary_op)) = result else {
         panic!("expected `BinaryOp`");
@@ -216,13 +182,10 @@ fn not_le_becomes_gt() {
 
 #[test]
 fn not_and_becomes_or_of_nots() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(a and b)` → `not(a) or not(b)`
     let mut expr = not_expr(Expr::and(Expr::arg(0), Expr::arg(1)));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::Or(or_expr)) = result else {
         panic!("expected `Or`");
@@ -234,13 +197,10 @@ fn not_and_becomes_or_of_nots() {
 
 #[test]
 fn not_or_becomes_and_of_nots() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(a or b)` → `not(a) and not(b)`
     let mut expr = not_expr(Expr::or(Expr::arg(0), Expr::arg(1)));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::And(and_expr)) = result else {
         panic!("expected `And`");
@@ -252,9 +212,6 @@ fn not_or_becomes_and_of_nots() {
 
 #[test]
 fn not_and_with_three_operands() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(a and b and c)` → `not(a) or not(b) or not(c)`
     let mut expr = not_expr(Expr::and_from_vec(vec![
         Expr::arg(0),
@@ -262,7 +219,7 @@ fn not_and_with_three_operands() {
         Expr::arg(2),
     ]));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::Or(or_expr)) = result else {
         panic!("expected `Or`");
@@ -275,9 +232,6 @@ fn not_and_with_three_operands() {
 
 #[test]
 fn not_or_with_three_operands() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(a or b or c)` → `not(a) and not(b) and not(c)`
     let mut expr = not_expr(Expr::or_from_vec(vec![
         Expr::arg(0),
@@ -285,7 +239,7 @@ fn not_or_with_three_operands() {
         Expr::arg(2),
     ]));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     let Some(Expr::And(and_expr)) = result else {
         panic!("expected `And`");
@@ -300,14 +254,11 @@ fn not_or_with_three_operands() {
 
 #[test]
 fn not_in_empty_list_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(x in ())` → `true`
     let in_list_expr = Expr::in_list(Expr::arg(0), Expr::list::<Expr>(vec![]));
     let mut expr = not_expr(in_list_expr);
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     assert!(result.is_some());
     assert!(result.unwrap().is_true());
@@ -315,17 +266,14 @@ fn not_in_empty_list_becomes_true() {
 
 #[test]
 fn not_in_non_empty_list_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `not(x in (1, 2))` → not simplified directly by expr_not
+    // `not(x in (1, 2))` → not simplified directly by fold_expr_not
     let in_list_expr = Expr::in_list(
         Expr::arg(0),
         Expr::list(vec![Value::from(1i64), Value::from(2i64)]),
     );
     let mut expr = not_expr(in_list_expr);
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     assert!(result.is_none());
 }
@@ -334,13 +282,10 @@ fn not_in_non_empty_list_not_simplified() {
 
 #[test]
 fn not_error_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `not(error("boom"))` → not simplified (error is not a value/binary_op/and/or/in_list)
     let mut expr = not_expr(Expr::error("boom"));
 
-    let result = simplify.simplify_expr_not(&mut expr);
+    let result = fold_expr_not(&mut expr);
 
     assert!(result.is_none());
 }
