+++ crates/toasty/src/engine/fold/tests/expr_record.rs
@@ -1,12 +1,8 @@
-use super::test_schema;
-use crate::engine::simplify::Simplify;
+use crate::engine::fold::expr_record::fold_expr_record;
 use toasty_core::stmt::{Expr, ExprRecord, Value};
 
 #[test]
 fn all_const_values_become_value_record() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `record([1, "hello", true]) → {1, "hello", true}`
     let mut expr = ExprRecord {
         fields: vec![
@@ -15,7 +11,7 @@ fn all_const_values_become_value_record() {
             Expr::Value(Value::from(true)),
         ],
     };
-    let result = simplify.simplify_expr_record(&mut expr);
+    let result = fold_expr_record(&mut expr);
 
     assert!(result.is_some());
     let Expr::Value(Value::Record(record)) = result.unwrap() else {
@@ -29,12 +25,9 @@ fn all_const_values_become_value_record() {
 
 #[test]
 fn empty_record_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `record([]) → {}`
     let mut expr = ExprRecord { fields: vec![] };
-    let result = simplify.simplify_expr_record(&mut expr);
+    let result = fold_expr_record(&mut expr);
 
     assert!(result.is_some());
     let Expr::Value(Value::Record(record)) = result.unwrap() else {
@@ -45,28 +38,22 @@ fn empty_record_simplified() {
 
 #[test]
 fn non_const_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `record([1, arg(0)])`, non-constant, not simplified
     let mut expr = ExprRecord {
         fields: vec![Expr::Value(Value::from(1i64)), Expr::arg(0)],
     };
-    let result = simplify.simplify_expr_record(&mut expr);
+    let result = fold_expr_record(&mut expr);
 
     assert!(result.is_none());
 }
 
 #[test]
 fn record_with_error_field_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
     // `record([1, error("boom")])` — error is not a Value, so not folded
     let mut expr = ExprRecord {
         fields: vec![Expr::Value(Value::from(1i64)), Expr::error("boom")],
     };
-    let result = simplify.simplify_expr_record(&mut expr);
+    let result = fold_expr_record(&mut expr);
 
     assert!(result.is_none());
 }
