+++ crates/toasty/src/engine/simplify/tests/expr_binary_op.rs
@@ -44,244 +44,6 @@ fn non_id_cast_not_unwrapped() {
 }
 
 #[test]
-fn constant_eq_same_values_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `eq(5, 5)` → `true`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_eq_different_values_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `eq(1, 2)` → `false`
-    let mut lhs = Expr::Value(Value::from(1i64));
-    let mut rhs = Expr::Value(Value::from(2i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
-}
-
-#[test]
-fn constant_ne_same_values_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `ne(5, 5)` → `false`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ne, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
-}
-
-#[test]
-fn constant_ne_different_values_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `ne("abc", "def")` → `true`
-    let mut lhs = Expr::Value(Value::from("abc"));
-    let mut rhs = Expr::Value(Value::from("def"));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ne, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_eq_with_null_becomes_null() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `eq(null, 5)` → `null`
-    let mut lhs = Expr::Value(Value::Null);
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Null))));
-}
-
-#[test]
-fn constant_eq_null_with_null_becomes_null() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `eq(null, null)` → `null`
-    let mut lhs = Expr::Value(Value::Null);
-    let mut rhs = Expr::Value(Value::Null);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Null))));
-}
-
-#[test]
-fn constant_lt_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `1 < 5` → `true`
-    let mut lhs = Expr::Value(Value::from(1i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_lt_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 < 1` → `false`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::Value(Value::from(1i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
-}
-
-#[test]
-fn constant_le_equal_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 <= 5` → `true`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Le, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_le_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `10 <= 5` → `false`
-    let mut lhs = Expr::Value(Value::from(10i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Le, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
-}
-
-#[test]
-fn constant_gt_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `10 > 5` → `true`
-    let mut lhs = Expr::Value(Value::from(10i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Gt, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_gt_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `1 > 5` → `false`
-    let mut lhs = Expr::Value(Value::from(1i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Gt, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
-}
-
-#[test]
-fn constant_ge_equal_becomes_true() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 >= 5` → `true`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ge, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_ge_becomes_false() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `1 >= 5` → `false`
-    let mut lhs = Expr::Value(Value::from(1i64));
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ge, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(false)))));
-}
-
-#[test]
-fn constant_lt_string_lexicographic() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `"abc" < "def"` → `true` (lexicographic)
-    let mut lhs = Expr::Value(Value::from("abc"));
-    let mut rhs = Expr::Value(Value::from("def"));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Value(Value::Bool(true)))));
-}
-
-#[test]
-fn constant_lt_different_types_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 < "abc"` is not simplified (incompatible types)
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::Value(Value::from("abc"));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
-
-    assert!(result.is_none());
-}
-
-#[test]
-fn lt_with_non_constant_not_simplified() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `arg(0) < 5` is not simplified (non-constant lhs)
-    let mut lhs = Expr::arg(0);
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
-
-    assert!(result.is_none());
-}
-
-#[test]
 fn self_comparison_eq_non_nullable_becomes_true() {
     let schema = test_schema();
     let model = schema.app.model(User::id());
@@ -365,185 +127,6 @@ fn different_fields_not_simplified() {
     });
 
     let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(result.is_none());
-}
-
-#[test]
-fn x_eq_true_becomes_x() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `x = true` → `x`
-    let mut lhs = Expr::arg(0);
-    let mut rhs = Expr::Value(Value::Bool(true));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Arg(_))));
-}
-
-#[test]
-fn true_eq_x_becomes_x() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `true = x` → `x`
-    let mut lhs = Expr::Value(Value::Bool(true));
-    let mut rhs = Expr::arg(0);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Arg(_))));
-}
-
-#[test]
-fn x_eq_false_becomes_not_x() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `x = false` → `not(x)`
-    let mut lhs = Expr::arg(0);
-    let mut rhs = Expr::Value(Value::Bool(false));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Not(_))));
-}
-
-#[test]
-fn x_ne_true_becomes_not_x() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `x != true` → `not(x)`
-    let mut lhs = Expr::arg(0);
-    let mut rhs = Expr::Value(Value::Bool(true));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ne, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Not(_))));
-}
-
-#[test]
-fn x_ne_false_becomes_x() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `x != false` → `x`
-    let mut lhs = Expr::arg(0);
-    let mut rhs = Expr::Value(Value::Bool(false));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ne, &mut lhs, &mut rhs);
-
-    assert!(matches!(result, Some(Expr::Arg(_))));
-}
-
-#[test]
-fn canonicalize_eq_literal_on_left() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 = x` → `x = 5`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::arg(0);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Eq, &mut lhs, &mut rhs);
-
-    let Some(Expr::BinaryOp(binary_op)) = result else {
-        panic!("expected BinaryOp");
-    };
-    assert_eq!(binary_op.op, BinaryOp::Eq);
-    assert!(matches!(*binary_op.lhs, Expr::Arg(_)));
-    assert!(matches!(*binary_op.rhs, Expr::Value(Value::I64(5))));
-}
-
-#[test]
-fn canonicalize_lt_literal_on_left() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 < x` → `x > 5`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::arg(0);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
-
-    let Some(Expr::BinaryOp(binary_op)) = result else {
-        panic!("expected BinaryOp");
-    };
-    assert_eq!(binary_op.op, BinaryOp::Gt);
-    assert!(matches!(*binary_op.lhs, Expr::Arg(_)));
-    assert!(matches!(*binary_op.rhs, Expr::Value(Value::I64(5))));
-}
-
-#[test]
-fn canonicalize_gt_literal_on_left() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 > x` → `x < 5`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::arg(0);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Gt, &mut lhs, &mut rhs);
-
-    let Some(Expr::BinaryOp(binary_op)) = result else {
-        panic!("expected BinaryOp");
-    };
-    assert_eq!(binary_op.op, BinaryOp::Lt);
-    assert!(matches!(*binary_op.lhs, Expr::Arg(_)));
-    assert!(matches!(*binary_op.rhs, Expr::Value(Value::I64(5))));
-}
-
-#[test]
-fn canonicalize_le_literal_on_left() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 <= x` → `x >= 5`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::arg(0);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Le, &mut lhs, &mut rhs);
-
-    let Some(Expr::BinaryOp(binary_op)) = result else {
-        panic!("expected BinaryOp");
-    };
-    assert_eq!(binary_op.op, BinaryOp::Ge);
-    assert!(matches!(*binary_op.lhs, Expr::Arg(_)));
-    assert!(matches!(*binary_op.rhs, Expr::Value(Value::I64(5))));
-}
-
-#[test]
-fn canonicalize_ge_literal_on_left() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `5 >= x` → `x <= 5`
-    let mut lhs = Expr::Value(Value::from(5i64));
-    let mut rhs = Expr::arg(0);
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Ge, &mut lhs, &mut rhs);
-
-    let Some(Expr::BinaryOp(binary_op)) = result else {
-        panic!("expected BinaryOp");
-    };
-    assert_eq!(binary_op.op, BinaryOp::Le);
-    assert!(matches!(*binary_op.lhs, Expr::Arg(_)));
-    assert!(matches!(*binary_op.rhs, Expr::Value(Value::I64(5))));
-}
-
-#[test]
-fn no_canonicalize_when_literal_on_right() {
-    let schema = test_schema();
-    let mut simplify = Simplify::new(&schema);
-
-    // `x < 5` is already canonical, no change
-    let mut lhs = Expr::arg(0);
-    let mut rhs = Expr::Value(Value::from(5i64));
-
-    let result = simplify.simplify_expr_binary_op(BinaryOp::Lt, &mut lhs, &mut rhs);
 
     assert!(result.is_none());
 }
